Finish first pass refactor
