+++ client/platform/web-girder/store/Brand.ts
@@ -16,9 +16,9 @@ function setFavicon(href: string) {
   faviconLink.setAttribute('href', href);
 }
 
-function setTitle(title: string) {
+function setTitle(title?: string) {
   const titleEl = document.querySelector('title');
-  if (titleEl) {
+  if (titleEl && title !== undefined) {
     titleEl.innerText = title;
   }
 }
@@ -28,7 +28,7 @@ const brandModule: Module<BrandState, RootState> = {
   state: {
     brandData: {
       vuetify: null,
-      favicon: null,
+      favicon: undefined,
       logo: defaultLogo,
       name: 'DIVE',
       loginMessage: `DIVE is automatically updated 
@@ -40,7 +40,7 @@ const brandModule: Module<BrandState, RootState> = {
     setBrandData(state, data: BrandData) {
       state.brandData = merge(state.brandData, data);
       setTitle(state.brandData.name);
-      if (state.brandData.favicon !== null) {
+      if (state.brandData.favicon) {
         setFavicon(state.brandData.favicon);
       }
     },
