+++ server/bucket_notifications/models.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from dive_server.utils import PydanticModel
+from dive_server.crud import PydanticModel
 
 
 class GCSObjectFinalizeNotification(BaseModel):
