+++ server/dive_server/__init__.py
@@ -5,41 +5,32 @@ from girder import events, plugin
 from girder.constants import AccessType
 from girder.models.setting import Setting
 from girder.models.user import User
-from girder.utility import mail_utils, setting_utilities
+from girder.utility import mail_utils
 from girder.utility.model_importer import ModelImporter
 
-from dive_utils.constants import SETTINGS_CONST_JOBS_CONFIGS, UserPrivateQueueEnabledMarker
+from dive_utils.constants import UserPrivateQueueEnabledMarker
 
-from .annotation_views import AnnotationResource
 from .client_webroot import ClientWebroot
-from .dataset_views import DatasetResource
 from .event import process_fs_import, process_s3_import, send_new_user_email
-from .viame import Viame
-from .viame_summary import SummaryItem, ViameSummary
-
-
-@setting_utilities.validator({SETTINGS_CONST_JOBS_CONFIGS})
-def validateSettings(doc):
-    """
-    Handle plugin-specific system settings. Right now we don't do any
-    validation.
-    """
-    val = doc['value']
-    if val is not None:
-        # TODO: replace with real schema validation
-        assert 'training' in val, '"training" missing from doc'
-        assert 'pipelines' in val, '"piplines" missing from doc'
+from .views_annotation import AnnotationResource
+from .views_dataset import DatasetResource
+from .views_override import use_private_queue
+from .views_rpc import RpcResource
+from .views_summary import SummaryItem, SummaryResource
 
 
 class GirderPlugin(plugin.GirderPlugin):
     def load(self, info):
         ModelImporter.registerModel('summaryItem', SummaryItem, plugin='dive_server')
-        User().exposeFields(AccessType.READ, UserPrivateQueueEnabledMarker)
 
-        info["apiRoot"].viame = Viame()
-        info["apiRoot"].viame_summary = ViameSummary()
+        info["apiRoot"].dive_summary = SummaryResource("dive_summary")
         info["apiRoot"].dive_annotation = AnnotationResource("dive_annotation")
         info["apiRoot"].dive_dataset = DatasetResource("dive_dataset")
+        info["apiRoot"].dive_rpc = RpcResource("dive_rpc")
+
+        # Setup route additions for exsting resources
+        info["apiRoot"].user.route("GET", (":id", "use_private_queue"), use_private_queue)
+        User().exposeFields(AccessType.READ, UserPrivateQueueEnabledMarker)
 
         DIVE_MAIL_TEMPLATES = Path(os.path.realpath(__file__)).parent / 'mail_templates'
         mail_utils.addTemplateDirectory(str(DIVE_MAIL_TEMPLATES))
