+++ server/dive_server/crud_annotation.py
@@ -1,19 +1,16 @@
-"""
-CRUD operations for the annotation REST view
-"""
 from typing import List
 
 from girder.exceptions import RestException
 from girder.models.file import File
 from pydantic.main import BaseModel
 
-from dive_server import utils
+from dive_server import crud
 from dive_utils import models, types
 
 
 def get_annotations(dsFolder: types.GirderModel):
-    utils.verify_dataset(dsFolder)
-    file = utils.detections_file(dsFolder)
+    crud.verify_dataset(dsFolder)
+    file = crud.detections_file(dsFolder)
     if file is None:
         return {}
     if "csv" in file["exts"]:
@@ -30,9 +27,9 @@ class AnnotationUpdateArgs(BaseModel):
 
 
 def save_annotations(dsFolder: types.GirderModel, user: types.GirderModel, data: dict):
-    utils.verify_dataset(dsFolder)
-    track_dict = utils.getTrackData(utils.detections_file(dsFolder))
-    validated: AnnotationUpdateArgs = utils.get_validated_model(AnnotationUpdateArgs, **data)
+    crud.verify_dataset(dsFolder)
+    track_dict = crud.getTrackData(crud.detections_file(dsFolder))
+    validated: AnnotationUpdateArgs = crud.get_validated_model(AnnotationUpdateArgs, **data)
 
     for track_id in validated.delete:
         track_dict.pop(str(track_id), None)
@@ -43,7 +40,7 @@ def save_annotations(dsFolder: types.GirderModel, user: types.GirderModel, data:
     deleted_len = len(validated.upsert)
 
     if upserted_len or deleted_len:
-        utils.saveTracks(dsFolder, track_dict, user)
+        crud.saveTracks(dsFolder, track_dict, user)
 
     return {
         "updated": upserted_len,
