+++ server/dive_server/views_dataset.py
@@ -6,7 +6,9 @@ from girder.api.rest import Resource, setContentDisposition, setResponseHeader
 from girder.constants import AccessType, TokenScope
 from girder.models.folder import Folder
 
-from . import dataset_crud
+from dive_utils import constants
+
+from . import crud_dataset
 
 DatasetModelParam = {
     'description': "dataset id",
@@ -17,29 +19,87 @@ DatasetModelParam = {
 
 
 class DatasetResource(Resource):
-    """
-    RESTful Dataset resource
-    """
+    """RESTful Dataset resource"""
 
     def __init__(self, resourceName):
         super(DatasetResource, self).__init__()
         self.resourceName = resourceName
 
+        self.route("POST", (), self.create_dataset)
+
+        self.route("GET", (), self.list_datasets)
         self.route("GET", (":id",), self.get_meta)
         self.route("GET", (":id", "export"), self.export)
+        self.route("GET", ("validate_files",), self.validate_files)
 
         self.route("PATCH", (":id", "metadata"), self.patch_metadata)
         self.route("PATCH", (":id", "attributes"), self.patch_attributes)
 
     @access.user
     @autoDescribeRoute(
+        Description("Create a new dataset")
+        .modelParam(
+            "cloneId",
+            description="Create dataset from clone source",
+            paramType="query",
+            destName="cloneSource",
+            model=Folder,
+            level=AccessType.READ,
+            required=True,
+        )
+        .modelParam(
+            "parentFolderId",
+            description="Parent folder",
+            paramType="query",
+            destName="parentFolder",
+            model=Folder,
+            level=AccessType.WRITE,
+            required=True,
+        )
+        .param(
+            "name",
+            "Name for new dataset",
+            paramType="query",
+            dataType="string",
+            default=None,
+            required=False,
+        )
+    )
+    def create_dataset(self, cloneSource, parentFolder, name):
+        # TODO: make this endpoint do regular creation and clone
+        return crud_dataset.createSoftClone(self.getCurrentUser(), cloneSource, parentFolder, name)
+
+    @access.user
+    @autoDescribeRoute(
+        Description("List datasets in the system")
+        .pagingParams("created")
+        .param(
+            constants.PublishedMarker,
+            'Return only published datasets',
+            required=False,
+            default=False,
+            dataType='boolean',
+        )
+    )
+    def list_datasets(self, params):
+        limit, offset, sort = self.getPagingParameters(params)
+        return crud_dataset.list_datasets(
+            self.getCurrentUser(),
+            self.boolParam(constants.PublishedMarker, params),
+            limit,
+            offset,
+            sort,
+        )
+
+    @access.user
+    @autoDescribeRoute(
         Description("Get dataset metadata").modelParam(
             "id", level=AccessType.READ, **DatasetModelParam
         )
+        # TODO add a "camera" query param
     )
     def get_meta(self, folder):
-        # TODO idea -- add a "camera" query param
-        return dataset_crud.get_dataset(folder, self.getCurrentUser()).dict(exclude_none=True)
+        return crud_dataset.get_dataset(folder, self.getCurrentUser()).dict(exclude_none=True)
 
     @access.public(scope=TokenScope.DATA_READ, cookie=True)
     @autoDescribeRoute(
@@ -85,7 +145,7 @@ class DatasetResource(Resource):
     ):
         setResponseHeader('Content-Type', 'application/zip')
         setContentDisposition(folder['name'] + '.zip')
-        return dataset_crud.export_dataset_zipstream(
+        return crud_dataset.export_dataset_zipstream(
             folder,
             self.getCurrentUser(),
             includeMedia=includeMedia,
@@ -93,6 +153,15 @@ class DatasetResource(Resource):
             excludeBelowThreshold=excludeBelowThreshold,
             typeFilter=typeFilter,
         )
+
+    @access.user
+    @autoDescribeRoute(
+        Description("Test whether or not a set of files are safe to upload").jsonParam(
+            "files", "", paramType="body", requireArray=True
+        )
+    )
+    def validate_files(self, files):
+        return crud_dataset.validate_files(files)
 
     @access.user
     @autoDescribeRoute(
@@ -106,7 +175,7 @@ class DatasetResource(Resource):
         )
     )
     def patch_metadata(self, folder, data):
-        return dataset_crud.update_dataset(folder, data)
+        return crud_dataset.update_dataset(folder, data)
 
     @access.user
     @autoDescribeRoute(
@@ -120,4 +189,4 @@ class DatasetResource(Resource):
         )
     )
     def patch_attributes(self, folder, data):
-        return dataset_crud.update_attributes(folder, data)
+        return crud_dataset.update_attributes(folder, data)
