+++ server/dive_server/views_rpc.py
@@ -1,0 +1,95 @@
+from girder.api import access
+from girder.api.describe import Description, autoDescribeRoute
+from girder.api.rest import Resource
+from girder.constants import AccessType
+from girder.models.folder import Folder
+from girder.models.token import Token
+
+from dive_utils.types import PipelineDescription
+
+from . import crud_rpc
+
+
+class RpcResource(Resource):
+    """Remote procedure calls to celery and other non-RESTful operations"""
+
+    def __init__(self, resourceName):
+        super(RpcResource, self).__init__()
+        self.resourceName = resourceName
+
+        self.route("POST", ("pipeline",), self.run_pipeline_task)
+        self.route("POST", ("train",), self.run_training)
+        self.route("POST", ("postprocess", ":id"), self.postprocess)
+
+    @access.user
+    @autoDescribeRoute(
+        Description("Run viame pipeline")
+        .modelParam(
+            "folderId",
+            description="Folder id of a video clip",
+            model=Folder,
+            paramType="query",
+            required=True,
+            level=AccessType.WRITE,
+        )
+        .jsonParam("pipeline", "The pipeline to run on the dataset", required=True)
+    )
+    def run_pipeline_task(self, folder, pipeline: PipelineDescription):
+        return crud_rpc.run_pipeline(self.getCurrentUser(), folder, pipeline)
+
+    @access.user
+    @autoDescribeRoute(
+        Description("Run training on a folder")
+        .jsonParam(
+            "folderIds",
+            description="Array of folderIds to run training on",
+            paramType="body",
+        )
+        .param(
+            "pipelineName",
+            description="The name of the resulting pipeline",
+            paramType="query",
+            required=True,
+        )
+        .param(
+            "config",
+            description="The configuration to use for training",
+            paramType="query",
+            required=True,
+        )
+        .param(
+            "annotatedFramesOnly",
+            description="Train only using frames with annotations",
+            paramType="query",
+            dataType="boolean",
+            default=False,
+            required=False,
+        )
+    )
+    def run_training(self, folderIds, pipelineName, config, annotatedFramesOnly):
+        user = self.getCurrentUser()
+        token = Token().createToken(user=user, days=14)
+        return crud_rpc.run_training(
+            user, token, folderIds, pipelineName, config, annotatedFramesOnly
+        )
+
+    @access.user
+    @autoDescribeRoute(
+        Description("Post-processing to be run after media/annotation import")
+        .modelParam(
+            "id",
+            description="Folder containing the items to process",
+            model=Folder,
+            level=AccessType.WRITE,
+        )
+        .param(
+            "skipJobs",
+            "Whether to skip processing that might dispatch worker jobs",
+            paramType="formData",
+            dataType="boolean",
+            default=False,
+            required=False,
+        )
+    )
+    def postprocess(self, folder, skipJobs):
+        return crud_rpc.postprocess(self.getCurrentUser(), folder, skipJobs)
